tagging version 1.2
git-svn-id: https://plugins.svn.wordpress.org/ios-images-fixer/trunk@1023091 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ image-fixer.php
@@ -3,7 +3,7 @@
  * Plugin Name: iOS Images Fixer
  * Plugin URI: http://bishoy.me/wp-plugins/ios-images-fixer/
  * Description: This plugin fixes iOS-taken images' orientation upon uploading using ImageMagic Library if available or PHP GD as a fallback. No settings editing required, just activate the plugin and try uploading an image from your idevice! If you like this free plugin, please <a href="http://bishoy.me/donate" target="_blank">consider a donation</a>.
- * Version: 1.1
+ * Version: 1.2
  * Author: Bishoy A.
  * Author URI: http://bishoy.me
  * License: GPL2
@@ -51,9 +51,48 @@ class BAImageFixer {
 	 * Start the plugin
 	 * @since 1.0
 	 */
-	public static function start() {
+	public function start() {
 		add_filter( 'wp_handle_upload_prefilter', array( self::get_instance(), 'imf_exif_rotate' ) );
-		add_action( 'admin_notices', array( self::get_instance(), 'required_function_notice' ) );
+		if ( is_admin() ){ 
+			add_action( 'admin_notices', array( self::get_instance(), 'required_function_notice' ) );
+			add_action( 'admin_menu', array( self::get_instance(), 'imf_menu' ) );
+		}
+	}
+
+	/**
+	 * Add Media Menu Page
+	 * @since  1.2
+	 * @return void
+	 */
+	public function imf_menu() {
+		$fixer_page = add_media_page('Fix iOS images', 'Fix iOS images', 'manage_options', 'ios-images-fixer', array( self::get_instance(), 'ios_fixer_page' ) );
+		add_action( 'admin_head-' . $fixer_page, array( self::get_instance(), 'admin_head' ) );
+	}
+
+	/**
+	 * Custom Styles for the fixer page
+	 * @since  1.2
+	 * @return mixed
+	 */
+	public function admin_head() { ?>
+		<style>
+			.fix-link {
+				color: #006000;
+				font-weight: bold;
+			}
+
+			.fix-link:hover {
+				color: green;
+			}
+		</style>
+	<?php }
+
+	/**
+	 * Page function
+	 * @return void
+	 */
+	public static function ios_fixer_page() {
+		require_once 'ios-fixer-page.php';
 	}
 
 	/**
@@ -102,18 +141,39 @@ class BAImageFixer {
 	}
 
 	/**
-	 * Rotate images to the correct orientation
-	 * @param  array $file $_FILES array
-	 * @return array	   $_FILES array in the correct orientation
-	 * @since 1.0
+	 * Get broken images from Media
+	 * @since  1.2
+	 * @return array
 	 */
-	public static function imf_exif_orient_correction( $file ) {
+	public static function get_broken_images() {
+		$query_images_args = array(
+		    'post_type'      => 'attachment', 
+		    'post_mime_type' =>'image/jpeg', 
+		    'post_status'    => 'inherit', 
+		    'posts_per_page' => -1,
+		);
 
-		if ( $file['type'] != 'image/jpeg' ) {
-			return $file;
+		$query_images = new WP_Query( $query_images_args );
+
+		$broken_images = array();
+
+		foreach ( $query_images->posts as $image ) {
+			$path = get_attached_file( $image->ID );
+			if ( self::is_image_broken( $path ) ) {
+				$broken_images[] = $image;
+			}
 		}
+		return $broken_images;
+	}
 
-		$exif = read_exif_data( $file['tmp_name'] );
+	/**
+	 * Checks if the image needs to be corrected
+	 * @since  1.2
+	 * @param  string  $path_to_image
+	 * @return boolean 
+	 */
+	public static function is_image_broken( $path_to_image ) {
+		$exif = read_exif_data( $path_to_image );
 		$exif_orient = isset($exif['Orientation'])?$exif['Orientation']:0;
 		$rotateImage = 0;
 
@@ -129,23 +189,63 @@ class BAImageFixer {
 		}
 
 		if ( $rotateImage ) {
+			return $rotateImage;
+		} else {
+			return 0;
+		}
+	}
+
+	/**
+	 * Rotate images to the correct orientation
+	 * @param  array $file $_FILES array
+	 * @return array	   $_FILES array in the correct orientation
+	 * @since 1.0
+	 */
+	public static function imf_exif_orient_correction( $file ) {
+
+		if ( $file['type'] != 'image/jpeg' ) {
+			return $file;
+		}
+
+		self::fix_orientation( $file['tmp_name'] );
+
+		return $file;
+	}
+
+	/**
+	 * Fixes image orientation
+	 * @param  string $img_path
+	 * @return boolean
+	 */
+	public static function fix_orientation( $img_path ) {
+
+		if ( self::something_is_wrong() ) {
+			return false;
+		}
+		
+		@set_time_limit( 900 );
+		$rotateImage = self::is_image_broken( $img_path );
+
+		if ( $rotateImage ) {
 			if ( class_exists( 'Imagick' ) ) {
 				$imagick = new Imagick();
 				$ImagickPixel = new ImagickPixel();
-				$imagick->readImage( $file['tmp_name'] );
+				$imagick->readImage( $img_path );
 				$imagick->rotateImage( $ImagickPixel, $rotateImage );
 				$imagick->setImageOrientation( $imageOrientation );
-				$imagick->writeImage( $file['tmp_name'] );
+				$imagick->writeImage( $img_path );
 				$imagick->clear();
 				$imagick->destroy();
 			} else {
 				$rotateImage = -$rotateImage;
-				$source = imagecreatefromjpeg( $file['tmp_name'] );
+				$source = imagecreatefromjpeg( $img_path );
 				$rotate = imagerotate( $source, $rotateImage, 0 );
-				imagejpeg( $rotate, $file['tmp_name'] );
+				imagejpeg( $rotate, $img_path );
 			}
+			return true;
+		} else {
+			return false;
 		}
-		return $file;
 	}
 }
